Moved GenLayerAtumBiome into a separate public class.
